refactor(cli): extract telemetry collection from monitor.rs
Rollback trigger: test regression or CI check failure.
